+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Upgraded `central-publishing-maven-plugin` to `0.9.0`.
 - Moved `Reflection.findSubclass` method to `Classes`.
 - Added `Serializable` to `ObjectConverter` to allow serialization of converters.
+- Added `Nullables.Chain.thenNotNull` to chain null checks.
 
 ---